moved schemaTestBts to separate schema file
